Update README.md
The work utilizes the contemporary tools
of artificial intelligence such as Jupyter Notebook in data cleaning, data visualization, data
modelling, and model assessment. Both Logistic regression and Random forest models are
used, and the results are compared based on precision-recall, F1-score, and AUC-ROC. It is
crucial to use the results and findings of this study to argue for the roles of AI solutions in
solving real-life issues and exploring the overall impact and future developments of
healthcare technologies.